app.py now split to 4 scripts
